# docs/swagger_dump.json
{"openapi": "3.1.0", "info": {"title": "Ado\u00e7\u00e3o C\u00e3osciente", "version": "0.1.0"}, "paths": {"/ong/": {"get": {"tags": ["ongs_data"], "summary": "Read Ong", "operationId": "read_ong_ong__get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"JWTBearer": []}]}, "put": {"tags": ["ongs_data"], "summary": "Update Ong", "operationId": "update_ong_ong__put", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OngModel"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"JWTBearer": []}]}, "post": {"tags": ["ongs_data"], "summary": "Create Ong", "operationId": "create_ong_ong__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OngModel"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["ongs_data"], "summary": "Delete Ong", "operationId": "delete_ong_ong__delete", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"JWTBearer": []}]}}, "/ong/animals": {"get": {"tags": ["ongs_data"], "summary": "Read Ong Animals", "operationId": "read_ong_animals_ong_animals_get", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}, "security": [{"JWTBearer": []}]}}, "/animal/{animal_id}": {"get": {"tags": ["animals_data"], "summary": "Read Animal", "operationId": "read_animal_animal__animal_id__get", "security": [{"JWTBearer": []}], "parameters": [{"name": "animal_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Animal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "put": {"tags": ["animals_data"], "summary": "Update Animal", "operationId": "update_animal_animal__animal_id__put", "security": [{"JWTBearer": []}], "parameters": [{"name": "animal_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Animal Id"}}], "requestBody": {"required": true, "content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimalModel"}}}}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}, "delete": {"tags": ["animals_data"], "summary": "Delete Animal", "operationId": "delete_animal_animal__animal_id__delete", "security": [{"JWTBearer": []}], "parameters": [{"name": "animal_id", "in": "path", "required": true, "schema": {"type": "string", "title": "Animal Id"}}], "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/animal/": {"post": {"tags": ["animals_data"], "summary": "Create Animal", "operationId": "create_animal_animal__post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/AnimalModel"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}, "security": [{"JWTBearer": []}]}}, "/login": {"post": {"tags": ["authentication"], "summary": "Login", "operationId": "login_login_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/LoginModel"}}}, "required": true}, "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/register": {"post": {"tags": ["authentication"], "summary": "Register", "operationId": "register_register_post", "requestBody": {"content": {"application/json": {"schema": {"$ref": "#/components/schemas/OngModel"}}}, "required": true}, "responses": {"201": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}, "422": {"description": "Validation Error", "content": {"application/json": {"schema": {"$ref": "#/components/schemas/HTTPValidationError"}}}}}}}, "/refresh": {"post": {"tags": ["authentication"], "summary": "Refresh", "operationId": "refresh_refresh_post", "responses": {"200": {"description": "Successful Response", "content": {"application/json": {"schema": {}}}}}}}}, "components": {"schemas": {"AnimalModel": {"properties": {"ong": {"type": "string", "title": "Ong"}, "name": {"type": "string", "title": "Name"}, "type": {"type": "string", "title": "Type"}, "breed": {"type": "string", "title": "Breed"}, "height": {"type": "number", "title": "Height"}, "weight": {"type": "number", "title": "Weight"}, "special_needs": {"items": {"type": "string"}, "type": "array", "title": "Special Needs"}, "adoption_requirements": {"type": "string", "title": "Adoption Requirements"}, "photo": {"type": "string", "title": "Photo"}, "animals": {"items": {"type": "string"}, "type": "array", "title": "Animals"}, "adopter": {"type": "string", "title": "Adopter"}}, "type": "object", "required": ["ong", "name", "type", "breed", "height", "weight", "special_needs", "adoption_requirements", "photo", "animals", "adopter"], "title": "AnimalModel"}, "HTTPValidationError": {"properties": {"detail": {"items": {"$ref": "#/components/schemas/ValidationError"}, "type": "array", "title": "Detail"}}, "type": "object", "title": "HTTPValidationError"}, "LoginModel": {"properties": {"cnpj": {"type": "string", "title": "Cnpj"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["cnpj", "password"], "title": "LoginModel"}, "OngModel": {"properties": {"cnpj": {"type": "string", "title": "Cnpj"}, "name": {"type": "string", "title": "Name"}, "logo": {"type": "string", "title": "Logo"}, "city": {"type": "string", "title": "City"}, "state": {"type": "string", "title": "State"}, "phone": {"type": "string", "title": "Phone"}, "email": {"type": "string", "title": "Email"}, "mission": {"type": "string", "title": "Mission"}, "foundation": {"type": "string", "title": "Foundation"}, "description": {"type": "string", "title": "Description"}, "animals": {"items": {"type": "string"}, "type": "array", "title": "Animals"}, "password": {"type": "string", "title": "Password"}}, "type": "object", "required": ["cnpj", "name", "logo", "city", "state", "phone", "email", "mission", "foundation", "description", "animals", "password"], "title": "OngModel"}, "ValidationError": {"properties": {"loc": {"items": {"anyOf": [{"type": "string"}, {"type": "integer"}]}, "type": "array", "title": "Location"}, "msg": {"type": "string", "title": "Message"}, "type": {"type": "string", "title": "Error Type"}}, "type": "object", "required": ["loc", "msg", "type"], "title": "ValidationError"}}, "securitySchemes": {"JWTBearer": {"type": "http", "scheme": "bearer"}}}}